+++ ui/ozone/common/BUILD.gn
@@ -40,35 +40,8 @@ source_set("common") {
     "//third_party/mesa:osmesa",
   ]
 
-  visibility = [ "//ui/ozone/platform/*" ]
-}
-
-if (ozone_platform_gbm) {
-  source_set("linux") {
-    assert(ozone_platform_gbm)
-
-    sources = [
-      "linux/drm_util_linux.cc",
-      "linux/drm_util_linux.h",
-      "linux/gbm_buffer_linux.cc",
-      "linux/gbm_buffer_linux.h",
-      "linux/gbm_device_linux.cc",
-      "linux/gbm_device_linux.h",
-      "linux/overlay_plane.cc",
-      "linux/overlay_plane.h",
-      "linux/scanout_buffer.h",
-    ]
-
-    deps = [
-      "//base:base",
-      "//third_party/libdrm",
-      "//third_party/minigbm",
-      "//ui/gfx:buffer_types",
-      "//ui/gfx:memory_buffer",
-      "//ui/gfx/geometry:geometry",
-      "//ui/ozone:ozone_base",
-    ]
-
-    visibility = [ "//ui/ozone/platform/drm/*" ]
-  }
+  visibility = [
+    "//ui/ozone/platform/*",
+    "//ui/ozone/common/linux/*",
+  ]
 }
