+++ ui/ozone/common/linux/BUILD.gn
@@ -1,0 +1,35 @@
+# Copyright 2018 The Chromium Authors. All rights reserved.
+# Use of this source code is governed by a BSD-style license that can be
+# found in the LICENSE file.
+
+import("//build/config/ui.gni")
+import("//ui/ozone/ozone.gni")
+
+assert(ozone_platform_gbm)
+
+source_set("linux") {
+  sources = [
+    "drm_util_linux.cc",
+    "drm_util_linux.h",
+    "gbm_buffer_linux.cc",
+    "gbm_buffer_linux.h",
+    "gbm_device_linux.cc",
+    "gbm_device_linux.h",
+    "gbm_surfaceless_linux.cc",
+    "gbm_surfaceless_linux.h",
+    "overlay_plane.cc",
+    "overlay_plane.h",
+    "scanout_buffer.h",
+  ]
+
+  deps = [
+    "//base:base",
+    "//third_party/libdrm",
+    "//third_party/minigbm",
+    "//ui/gfx:buffer_types",
+    "//ui/gfx:memory_buffer",
+    "//ui/gfx/geometry:geometry",
+    "//ui/ozone:ozone_base",
+    "//ui/ozone/common",
+  ]
+}
