+++ ui/ozone/platform/drm/BUILD.gn
@@ -138,7 +138,7 @@ source_set("gbm") {
     "//ui/gl",
     "//ui/ozone:ozone_base",
     "//ui/ozone/common",
-    "//ui/ozone/common:linux",
+    "//ui/ozone/common/linux",
     "//ui/ozone/public/interfaces",
     "//ui/platform_window",
   ]
@@ -177,6 +177,6 @@ source_set("gbm_unittests") {
     "//ui/gfx",
     "//ui/ozone:platform",
     "//ui/ozone/common",
-    "//ui/ozone/common:linux",
+    "//ui/ozone/common/linux",
   ]
 }
